Wall Jump Functionality
Added a majority of the code needed to make the wall jump functional, for right now, there seems to be a bug where it won't consistently check with the raycast the player sends out all the time, and I need to fix it. Might also have to potentially fix a same-wall wall jump bug depending on how we want to design the game.