Bool Fixes
Updated all the enemy behavior scripts that were using the 'facingright' bool after it was added to the basic enemy so that there's no longer errors with it. Also fixed werewolf facing the wrong direction, only actively turning to face the player when chasing them and lastly, fixed the dash attack functionality to make it do damage properly.