Werewolf Completion
Although not perfect, I consider the werewolf at this point in time, completed. It has the correct spacing functionality as well as the ability to go between close and long range attacks. Also adjusted the run attack to make sure the wolf tries to close the distance first before doing their tun attack.