Enemy Code Cleanup and Functionality Revamp
cleaned up the patrol functionality set up on most enemies to be moved over into the base enemy, to reduce points of potential failure across multiple scripts. Also slightly altered the order in which the code is executed for patrolling for enemies, opting to set the target location once since everything else got moved to and handled in the base enemy. Also implemented ledge detection and prevented the enemy from walking off a ledge when moving towards a patrol point. Also found a couple of bugs and issues in some enemies that have since been fixed.